Trying to fix issue when you jump close to the end of the list.

+++ PockeTwit/Following.cs
@@ -101,7 +101,10 @@ namespace PockeTwit
             }
             catch
             {
-                System.IO.File.Delete(location);   
+                if (!string.IsNullOrEmpty(location))
+                {
+                    System.IO.File.Delete(location);
+                }
             }
             
         }
